new_requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/

ipykernel
jupyter
pandas
matplotlib
transformers
torch
paddlepaddle==3.0.0
paddleocr
pytesseract
pillow
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-core==0.1.52
python-dotenv
landingai-ade
pymupdf
pydantic_settings
chromadb
langchain-groq==0.1.3
sentence-transformers==2.7.0